+++ rocket-hal-driver/src/driver.rs
@@ -153,17 +153,16 @@ unsafe extern "C" fn factory_try_create(
     host_allocator: iree_allocator_t,
     out_driver: *mut *mut iree_hal_driver_t,
 ) -> iree_status_t {
-    let requested = unsafe {
-        std::slice::from_raw_parts(driver_name.data as *const u8, driver_name.size as usize)
-    };
+    let requested =
+        unsafe { std::slice::from_raw_parts(driver_name.data as *const u8, driver_name.size) };
     if requested != DRIVER_NAME {
-        return status::from_code(iree_status_code_e_IREE_STATUS_UNAVAILABLE as u32);
+        return status::from_code(iree_status_code_e_IREE_STATUS_UNAVAILABLE);
     }
 
     // Probe -- confirms the kernel driver is actually present rather than
     // deferring that discovery to the first create_device_by_id call.
     if std::fs::metadata(DEVICE_PATH).is_err() {
-        return status::from_code(iree_status_code_e_IREE_STATUS_UNAVAILABLE as u32);
+        return status::from_code(iree_status_code_e_IREE_STATUS_UNAVAILABLE);
     }
 
     let driver = Box::new(RocketDriver {
